Updated nulls handling (#590)

+++ components/vector/data_chunk.hpp
@@ -39,8 +39,8 @@ namespace components::vector {
         void set_value(const std::pmr::vector<size_t>& col_path, uint64_t index, const types::logical_value_t& val);
 
         bool is_null(uint64_t col_idx, uint64_t index) const;
-        bool is_null(uint64_t col_idx, const std::pmr::vector<uint64_t>& path) const;
-        void set_null(uint64_t col_idx, const std::pmr::vector<uint64_t>& path, bool value);
+        bool is_null(uint64_t col_idx, const std::pmr::vector<size_t>& path) const;
+        void set_null(uint64_t col_idx, const std::pmr::vector<size_t>& path, bool value);
         template<typename T>
         T get_value(uint64_t col_idx, uint64_t index) const {
             return data[col_idx].get_value<T>(index);
@@ -121,7 +121,8 @@ namespace components::vector {
     // Sub-query result compacters: fold ALL of a cursor's chunks (not just the first) into one bound
     // parameter value, so a >1024-row or multi-branch result is not silently truncated.
     core::result_wrapper_t<types::logical_value_t> compact_to_bool_value(const std::pmr::vector<data_chunk_t>& chunks);
-    core::result_wrapper_t<types::logical_value_t> compact_to_single_value(const std::pmr::vector<data_chunk_t>& chunks);
+    core::result_wrapper_t<types::logical_value_t>
+    compact_to_single_value(const std::pmr::vector<data_chunk_t>& chunks);
     core::result_wrapper_t<types::logical_value_t> compact_to_array_value(const std::pmr::vector<data_chunk_t>& chunks);
     core::result_wrapper_t<types::logical_value_t> compact_to_row_value(const std::pmr::vector<data_chunk_t>& chunks);
